harvard-cs-2881-hw1-RL/pyproject.toml
[project]
name = "hw1-rl-prompt-prefix"
version = "0.1.0"
description = "HW1: Prompt Prefix Optimization via Policy Gradients"
readme = "README_EXPERIMENT.md"
requires-python = ">=3.8"
dependencies = [
    # Core dependencies
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    # Deep Learning & Transformers
    "torch>=1.12.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    # Datasets and benchmarks
    "datasets>=2.0.0",
    # Utilities
    "tqdm>=4.0.0",
    "python-dotenv>=1.0.1",
    # Optional: Jupyter notebook support
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
quantization = [
    "bitsandbytes>=0.41.0; platform_system == 'Linux'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
